Moving bits of LBS solver into the LBS solver directory
Closes #41
